Move the adaptive sampler to a shared location. The rate limiting sampler is not specific to exception profiling only so it should live in a more generic space.
